Move waiter test over to waiter integ test
